separate env code into a new module
