+++ src/poke_env/battle/battle.py
@@ -1,9 +1,9 @@
 from logging import Logger
 from typing import Any, Dict, List, Optional, Union
 
-from poke_env.environment.abstract_battle import AbstractBattle
-from poke_env.environment.move import Move
-from poke_env.environment.pokemon import Pokemon
+from poke_env.battle.abstract_battle import AbstractBattle
+from poke_env.battle.move import Move
+from poke_env.battle.pokemon import Pokemon
 
 
 class Battle(AbstractBattle):
