+++ src/poke_env/battle/double_battle.py
@@ -1,12 +1,12 @@
 from logging import Logger
 from typing import Any, Dict, List, Optional, Union
 
-from poke_env.environment.abstract_battle import AbstractBattle
-from poke_env.environment.move import SPECIAL_MOVES, Move
-from poke_env.environment.move_category import MoveCategory
-from poke_env.environment.pokemon import Pokemon
-from poke_env.environment.pokemon_type import PokemonType
-from poke_env.environment.target import Target
+from poke_env.battle.abstract_battle import AbstractBattle
+from poke_env.battle.move import SPECIAL_MOVES, Move
+from poke_env.battle.move_category import MoveCategory
+from poke_env.battle.pokemon import Pokemon
+from poke_env.battle.pokemon_type import PokemonType
+from poke_env.battle.target import Target
 
 
 class DoubleBattle(AbstractBattle):
