+++ src/poke_env/battle/observation.py
@@ -8,10 +8,10 @@ with the Observations' properties, and then recreate that turn with the events p
 from dataclasses import dataclass, field
 from typing import Dict, List, Optional, Union
 
-from poke_env.environment.field import Field
-from poke_env.environment.observed_pokemon import ObservedPokemon
-from poke_env.environment.side_condition import SideCondition
-from poke_env.environment.weather import Weather
+from poke_env.battle.field import Field
+from poke_env.battle.observed_pokemon import ObservedPokemon
+from poke_env.battle.side_condition import SideCondition
+from poke_env.battle.weather import Weather
 
 
 @dataclass
