+++ src/poke_env/battle/z_crystal.py
@@ -4,7 +4,7 @@ directly.
 
 from typing import Dict, Optional, Tuple
 
-from poke_env.environment.pokemon_type import PokemonType
+from poke_env.battle.pokemon_type import PokemonType
 
 Z_CRYSTAL: Dict[str, Tuple[Optional[PokemonType], Optional[str]]] = {
     "buginiumz": (PokemonType.BUG, None),
