+++ src/poke_env/environment/env.py
@@ -14,13 +14,13 @@ from gymnasium.utils import seeding
 from numpy.random import Generator
 from pettingzoo.utils.env import ParallelEnv  # type: ignore[import-untyped]
 
-from poke_env.concurrency import POKE_LOOP, create_in_poke_loop
-from poke_env.environment.abstract_battle import AbstractBattle
-from poke_env.player.battle_order import (
+from poke_env.battle.abstract_battle import AbstractBattle
+from poke_env.battle.battle_order import (
     BattleOrder,
     DefaultBattleOrder,
     ForfeitBattleOrder,
 )
+from poke_env.concurrency import POKE_LOOP, create_in_poke_loop
 from poke_env.player.player import Player
 from poke_env.ps_client import AccountConfiguration
 from poke_env.ps_client.server_configuration import (
