+++ src/poke_env/environment/single_agent_wrapper.py
@@ -2,7 +2,7 @@ from typing import Any, Awaitable, Dict, Optional, Tuple
 
 from gymnasium import Env
 
-from poke_env.player.env import ActionType, ObsType, PokeEnv
+from poke_env.environment.env import ActionType, ObsType, PokeEnv
 from poke_env.player.player import Player
