+++ src/poke_env/player/__init__.py
@@ -1,23 +1,18 @@
 """poke_env.player module init."""
 
-from pettingzoo.utils.env import ActionType, ObsType  # type: ignore[import-untyped]
-
+from poke_env.battle.battle_order import (
+    BattleOrder,
+    DefaultBattleOrder,
+    DoubleBattleOrder,
+    ForfeitBattleOrder,
+)
 from poke_env.concurrency import POKE_LOOP
 from poke_env.player.baselines import (
     MaxBasePowerPlayer,
     RandomPlayer,
     SimpleHeuristicsPlayer,
 )
-from poke_env.player.battle_order import (
-    BattleOrder,
-    DefaultBattleOrder,
-    DoubleBattleOrder,
-    ForfeitBattleOrder,
-)
-from poke_env.player.env import PokeEnv
 from poke_env.player.player import Player
-from poke_env.player.single_agent_wrapper import SingleAgentWrapper
-from poke_env.player.singles_env import SinglesEnv
 from poke_env.player.utils import (
     background_cross_evaluate,
     background_evaluate_player,
@@ -27,13 +22,8 @@ from poke_env.player.utils import (
 from poke_env.ps_client import PSClient
 
 __all__ = [
-    "ActionType",
-    "ObsType",
     "ForfeitBattleOrder",
     "POKE_LOOP",
-    "PokeEnv",
-    "SinglesEnv",
-    "SingleAgentWrapper",
     "PSClient",
     "Player",
     "cross_evaluate",
