+++ src/poke_env/player/baselines.py
@@ -1,17 +1,17 @@
 import random
 from typing import List, Optional
 
-from poke_env.environment.abstract_battle import AbstractBattle
-from poke_env.environment.double_battle import DoubleBattle
-from poke_env.environment.move_category import MoveCategory
-from poke_env.environment.pokemon import Pokemon
-from poke_env.environment.side_condition import SideCondition
-from poke_env.environment.target import Target
-from poke_env.player.battle_order import (
+from poke_env.battle.abstract_battle import AbstractBattle
+from poke_env.battle.battle_order import (
     BattleOrder,
     DefaultBattleOrder,
     DoubleBattleOrder,
 )
+from poke_env.battle.double_battle import DoubleBattle
+from poke_env.battle.move_category import MoveCategory
+from poke_env.battle.pokemon import Pokemon
+from poke_env.battle.side_condition import SideCondition
+from poke_env.battle.target import Target
 from poke_env.player.player import Player
